+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change History
 
+# 3.0
+
+- @tbela99 #138 AVIF image support: automatically convert images to AVIF format when the browser advertise it and the dg support is enabled
+- @tbela99 #84 Automatically generate critical css path. no more parameter are required
+
 # 2.11.0
 
 - @cbahiana-sd Added info on enabling zlib.output_compression 495b84d